fix(core): can not preview image in shared page (#7466)
FIx BS-798
